Added converstion from string to int[] function
Added a function to convert from a string of space separated integers
to an array of those values. I updated the setRow function to use
this new function.

This was done with the goal of making matrixCalc.java only do IO and
leave all real calculation/manipulation to Matrix.java and its functions.